+++ logger.go
@@ -1,0 +1,8 @@
+package pirsch
+
+import (
+	"log"
+	"os"
+)
+
+var logger = log.New(os.Stdout, "[pirsch] ", log.LstdFlags)